Refactor. put login.jsx into page folder
