+++ src/components/app/App.jsx
@@ -1,5 +1,5 @@
 import { Routes, Route, Navigate } from "react-router-dom";
-import Login from "../Login";
+import Login from "../pages/Login";
 
 function App() {
   return (
